# rest-api/docs/swagger.yaml
basePath: /v1
definitions:
...
  post.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  post.PostResponse:
    properties:
      post:
        $ref: '#/definitions/post.PostSwagger'
    type: object
  post.PostSwagger:
    type: object
  post.PostsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/post.PostSwagger'
        type: array
      total:
        type: integer
    type: object
  post.UnauthorizedResponse:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  role.AssignRoleRequest:
    properties:
      role:
        type: string
    type: object
  role.ErrorRoleResponse:
    properties:
      error:
        type: string
    type: object
  user.Public:
    properties:
      email:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  user.User:
    type: object
host: localhost:8080
info:
  contact: {}
  description: Example REST API with JWT Authentication
  title: REST API Example
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Get list of users with optional filtering
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get list of users
      tags:
      - Admin
  /admin/users/{userId}:
    get:
      description: Get user details by ID, or the logged-in user if ID not provided
      parameters:
      - description: User ID
        in: path
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Admin
  /admin/users/{userId}/assign:
    post:
      description: Assign a specific role (e.g. Writter) to a user by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Role to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.AssignRoleRequest'
      responses:
        "200":
          description: Role assigned successfully
        "400":
          description: Invalid role
          schema:
            $ref: '#/definitions/role.ErrorRoleResponse'
        "500":
          description: Failed to assign role
          schema:
            $ref: '#/definitions/role.ErrorRoleResponse'
      security:
      - BearerAuth: []
      summary: Assign a role to a user
      tags:
      - Admin
  /admin/users/{userId}/unassign:
    post:
      description: Remove any assigned role from a user by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: Role unassigned successfully
        "500":
          description: Failed to unassign role
          schema:
            $ref: '#/definitions/role.ErrorRoleResponse'
      security:
      - BearerAuth: []
      summary: Remove role from a user
      tags:
      - Admin
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get list of posts with optional search, filters, and pagination
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/post.ErrorResponse'
      summary: List posts
      tags:
      - Post
  /posts/{postID}:
    get:
      consumes:
      - application/json
      description: Get post details by postID
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/post.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/post.ErrorResponse'
      summary: Get a post by ID
      tags:
      - Post
...
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user in the system
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - User
  /users/{userId}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update fields of a user
      parameters:
      - description: User ID
        in: path
        name: userId
        type: integer
      - description: Updated user fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
  /users/me:
    get:
      description: Get logged-in user info based on JWT token
      produces:
      - application/json
      responses:
        "200":
          description: Current user info
          schema:
            $ref: '#/definitions/user.Public'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/post.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/post.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - User
  /writter/posts:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Post payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/post.PostSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/post.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/post.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a post
      tags:
      - Post
  /writter/posts/{postID}:
    delete:
      consumes:
      - application/json
      description: Soft delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/post.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/post.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/post.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - Post
    put:
      consumes:
      - application/json
      description: Update an existing post by ID
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      - description: Post payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/post.PostSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/post.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/post.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - Post
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"